Practice 12
Managing Tree components
