+++ src2/components/ToDoItem.jsx
@@ -1,8 +1,6 @@
-import React, { useState } from "react";
+import React from "react";
 
 function ToDoItem(props) {
-  function handleClick() {}
-
   return (
     <div
       onClick={() => {
